Allow saturation to be a LONG INT useful for pixels with very large values i.e. greater than 2^31

+++ iotbx/detectors/detectors_ext.cpp
@@ -372,7 +372,7 @@ struct flex_image_wrapper {
 
     class_<w_t >(python_name, no_init)
       .def(init<array_t&, const int&, const std::string&,
-                double const&, int const&, bool const& >(
+                double const&, long int const&, bool const& >(
             (
             arg_("rawdata"),
             arg_("binning"),
